Move model to it's own file
